pull setRelations and setTargetContent up (WP-854)
