+++ inc/Smartling/ContentTypes/Elementor/Elements/Unknown.php
@@ -58,52 +58,4 @@ class Unknown extends ElementAbstract {
         return ElementFactory::UNKNOWN_ELEMENT;
     }
 
-    public function setRelations(Content $content,
-        string $path,
-        SubmissionEntity $submission,
-        SubmissionManager $submissionManager,
-    ): Element
-    {
-        $result = clone $this;
-        $target = $submissionManager->findTargetBlogSubmission(
-            $content->getContentType(),
-            $submission->getSourceBlogId(),
-            $content->getContentId(),
-            $submission->getTargetBlogId(),
-        );
-        if ($target !== null) {
-            $result->raw = array_replace_recursive(
-                $result->raw,
-                (new ArrayHelper())->structurize([$path => $target->getTargetId()]),
-            );
-        }
-
-        return $result;
-    }
-
-    public function setTargetContent(
-        RelatedContentInfo $info,
-        array $strings,
-        SubmissionEntity $submission,
-        SubmissionManager $submissionManager,
-    ): Element
-    {
-        foreach ($this->elements as &$element) {
-            if ($element instanceof Element) {
-                $element = $element->setTargetContent(
-                    new RelatedContentInfo($info->getInfo()[$this->id] ?? []),
-                    $strings[$this->id][$element->id] ?? [],
-                    $submission,
-                    $submissionManager,
-                );
-            }
-        }
-        unset ($element);
-        foreach ($info->getOwnRelatedContent($this->id) as $path => $content) {
-            assert($content instanceof Content);
-            $this->raw['elements'] = $this->elements;
-            $this->raw = $this->setRelations($content, $path, $submission, $submissionManager)->toArray();
-        }
-        return new self($this->raw);
-    }
 }
